Removed backend fetch to a repo class.
